+++ CHANGELOG.md
@@ -19,7 +19,14 @@ ChangeLog
   REPORT (@DeepDiver1975).
 * #918: Add a lot of sqlite indexes. This should speed up sqlite-based
   installations quite a bit.
+* #982: Make sure that files that are siblings of directories, are reported
+  as files (@nickvergessen)
 
+3.2.3 (????-??-??)
+------------------
+
+* #982: Make sure that files that are siblings of directories, are reported
+  as files (@nickvergessen)
 
 3.2.2 (2017-02-14)
 ------------------